requirement.txt
flask
gunicorn
numpy
pillow
tensorflow